test/pyproject.toml
[tool.poetry]
name = "cicd_test_suite"
version = "0.1.0"
description = ""
authors = ["andreaba0 <andreabarchietto01@gmail.com>"]
readme = "README.md"
packages = [
    { include = "cicd_test_suite" }
]

[tool.poetry.dependencies]
python = "^3.10"
docker = "^7.0.0"
psycopg2-binary = "^2.9.9"
python-jose = "^3.3.0"
faker = "^23.1.0"
pycryptodome = "^3.20.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
